requirements.txt
streamlit
ultralytics
Pillow
numpy
requests
opencv-python